move verifier tests into a separate file;
